Make DXFReader restart after identifying file encoding
After reading the HEADER section, look for $ACADVER and $DWGCODEPAGE
variables, to identify the correct encoding for reading the file. If this
results in some encoding other than None, the entire read process is
restarted with the input file using the correct encoding.
This is needed because older AUtoCAD DXF files were not UTF-8 encoded,
but could contain cp1252 text that is invalid utf-8 causing errors or
incorrect text in the imported objects.